불필요

+++ core/src/main/kotlin/com/oop/game/world/ZombieWorld.kt
@@ -188,7 +188,7 @@ class ZombieWorld(game: ZombieGame, width: Float = game.screenWidth.toFloat(), h
 	private inline fun updateProgressBars() {
 		// HP 미터기 처리
 		val hpIndicator = getWidget("hp_indicator") as ProgressBar;
-		hpIndicator.value = player.hp.toFloat() / player.maxHp.toFloat();
+		hpIndicator.value = player.hp.toFloat() / player.maxHp;
 		
 		// 총 관련 미터기 처리
 		val ammoIndicator = getWidget("gun_ammo_indicator") as ProgressBar;
